requirements.txt
Django
django_registration
django-notifications-hq
django-crispy-forms
whitenoise
gunicorn